add helper class to combine repeated code
