+++ src/test/kotlin/uiController/LoginUIControllerTest.kt
@@ -1,29 +1,24 @@
 package uiController
 
+import helper.SUCCESS_MESSAGE_FOR_LOGIN
+import helper.USER_NOT_FOUND
+import helper.WRONG_PASSWORD
+import helper.WRONG_USER_NAME
+import helper.readInputFromConsole
+import helper.validUserData
 import io.mockk.every
 import io.mockk.mockk
-import logic.entities.User
 import logic.exception.PlanMateException.AuthorizationException.*
 import logic.usecases.LoginUseCase
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import ui.LoginUIController
-import java.io.ByteArrayInputStream
-import java.io.ByteArrayOutputStream
-import java.io.PrintStream
-import java.util.UUID
 import kotlin.test.Test
 
 class LoginUIControllerTest {
     private lateinit var useCase : LoginUseCase
     private lateinit var uiController : LoginUIController
-    private  val validAdminUserData = User(
-        username = "Hend",
-        password = "H123456",
-        isAdmin = true,
-        id = UUID.fromString("ebcb217c-b373-4e88-afbd-cbb5640a031a")
-    )
 
     @BeforeEach
     fun setup(){
@@ -32,20 +27,17 @@ class LoginUIControllerTest {
     }
 
     @Test
-    fun `should print success message when username and password are valid format and useCase return the role of this correct user`(){
+    fun `should print success message when username and password are valid format and useCase return full user data`(){
         val username = "Hend"
         val password = "H123456"
         val input = "$username\n$password"
-        System.setIn(ByteArrayInputStream(input.toByteArray()))
 
-        val outputStream = ByteArrayOutputStream()
-        System.setOut(PrintStream(outputStream))
+        val output = readInputFromConsole(input)
 
-        every { useCase.verifyUserState(username,password) } returns validAdminUserData.isAdmin
+        every { useCase.verifyUserState(username,password) } returns validUserData()
 
         uiController.loginUIController()
 
-        val output = outputStream.toString()
         assert(output.contains(SUCCESS_MESSAGE_FOR_LOGIN))
     }
 
@@ -54,16 +46,13 @@ class LoginUIControllerTest {
         val username = "nada"
         val password = "H12345"
         val input = "$username\n$password"
-        System.setIn(ByteArrayInputStream(input.toByteArray()))
 
-        val outputStream = ByteArrayOutputStream()
-        System.setOut(PrintStream(outputStream))
+        val output = readInputFromConsole(input)
 
         every { useCase.verifyUserState(username,password) } throws UserNotFoundException
 
         uiController.loginUIController()
 
-        val output = outputStream.toString()
         assert(output.contains(USER_NOT_FOUND))
     }
 
@@ -71,17 +60,13 @@ class LoginUIControllerTest {
     fun `should print wrong username message when username and password are valid format but useCase throw WrongUserName exception`(){
         val username = "Hen"
         val password = "H123456"
-        val input = "$username\n$password"
-        System.setIn(ByteArrayInputStream(input.toByteArray()))
 
-        val outputStream = ByteArrayOutputStream()
-        System.setOut(PrintStream(outputStream))
+        val output = readInputFromConsole(username)
 
         every { useCase.verifyUserState(username,password) } throws WrongUsername
 
         uiController.loginUIController()
 
-        val output = outputStream.toString()
         assert(output.contains(WRONG_USER_NAME))
     }
 
@@ -90,16 +75,13 @@ class LoginUIControllerTest {
         val username = "Hend"
         val password = "12345"
         val input = "$username\n$password"
-        System.setIn(ByteArrayInputStream(input.toByteArray()))
 
-        val outputStream = ByteArrayOutputStream()
-        System.setOut(PrintStream(outputStream))
+        val output = readInputFromConsole(input)
 
         every { useCase.verifyUserState(username,password) } throws WrongPassword
 
         uiController.loginUIController()
 
-        val output = outputStream.toString()
         assert(output.contains(WRONG_PASSWORD))
     }
 
@@ -114,44 +96,29 @@ class LoginUIControllerTest {
     )
     fun `should print wrong username message when username is invalid format`(username:String){
 
-        System.setIn(ByteArrayInputStream(username.toByteArray()))
-
-        val outputStream = ByteArrayOutputStream()
-        System.setOut(PrintStream(outputStream))
+        val output = readInputFromConsole(username)
 
         uiController.loginUIController()
 
-        val output = outputStream.toString()
         assert(output.contains(WRONG_USER_NAME))
     }
 
     @ParameterizedTest
     @ValueSource(
         strings = [
-            "123456", //only digits
+            "123456",
              "",
-            "h1234" //length less than 6
+            "h1234"
         ]
     )
     fun `should print wrong password message when password is invalid format`(password:String){
         val username = "Hend"
         val input = "$username\n$password"
-        System.setIn(ByteArrayInputStream(input.toByteArray()))
 
-        val outputStream = ByteArrayOutputStream()
-        System.setOut(PrintStream(outputStream))
+        val output = readInputFromConsole(input)
 
         uiController.loginUIController()
 
-        val output = outputStream.toString()
         assert(output.contains(WRONG_PASSWORD))
-    }
-
-
-    companion object{
-        const val SUCCESS_MESSAGE_FOR_LOGIN="Success Login"
-        const val USER_NOT_FOUND="This User Not Found"
-        const val WRONG_USER_NAME="This Username is Wrong , Please Enter Correct Name"
-        const val WRONG_PASSWORD="This Password is Wrong , Please Enter Correct Password"
     }
 }
