+++ src/test/kotlin/usecase/LoginUseCaseTest.kt
@@ -1,8 +1,8 @@
 package usecase
 
+import helper.validUserData
 import io.mockk.every
 import io.mockk.mockk
-import logic.entities.User
 import logic.exception.PlanMateException.AuthorizationException.*
 import logic.repository.AuthRepository
 import logic.usecases.LoginUseCase
@@ -10,24 +10,11 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import java.util.UUID
-import kotlin.test.assertFalse
 
 class LoginUseCaseTest {
     private lateinit var repository : AuthRepository
     private lateinit var useCase : LoginUseCase
-    private  val validAdminUserData = User(
-        username = "Hend",
-        password = "H123456",
-        isAdmin = true,
-        id = UUID.fromString("ebcb217c-b373-4e88-afbd-cbb5640a031a")
-    )
-    private  val validMateUserData = User(
-        username = "Nada",
-        password = "N12345",
-        isAdmin = false,
-        id = UUID.fromString("ebcb257c-b777-4r88-afbk-cbb5640a031a")
-    )
+
 
     @BeforeEach
     fun setup(){
@@ -36,39 +23,15 @@ class LoginUseCaseTest {
     }
 
     @Test
-    fun `should return role of user when username and password are found`(){
+    fun `should return full user data when username and password are found`(){
         //Given
         val username= "Hend"
         val password = "H123456"
         //When
-        every { repository.login(username, password) } returns validAdminUserData
+        every { repository.login(username, password) } returns validUserData()
         //then
-        val returnedRole = useCase.verifyUserState(username, password)
-        Assertions.assertEquals(validAdminUserData.isAdmin, returnedRole)
-    }
-
-    @Test
-    fun `should return role is true when username and password are valid for admin`(){
-        //Given
-        val username= "Hend"
-        val password = "H123456"
-        //When
-        every { repository.login(username, password) } returns validAdminUserData
-        //then
-        val returnedRole = useCase.verifyUserState(username, password)
-        Assertions.assertTrue(returnedRole)
-    }
-
-    @Test
-    fun `should return role is false when username and password are valid for mateUser`(){
-        //Given
-        val username= "Nada"
-        val password = "N12345"
-        //When
-        every { repository.login(username, password) } returns validMateUserData
-        //then
-        val returnedRole = useCase.verifyUserState(username, password)
-        assertFalse(returnedRole)
+        val returnedUser = useCase.verifyUserState(username, password)
+        Assertions.assertEquals(validUserData(), returnedUser)
     }
 
     @Test
@@ -105,11 +68,7 @@ class LoginUseCaseTest {
         val username= "Hen"
         val password = "K12345"
 
-        every {
-            repository.login(
-                username,
-                password
-            )
+        every { repository.login(username, password)
         } throws UserNotFoundException
 
         //When && Throw
